Update packagename to io.crossbar.autobahn
