+++ autobahn/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-      package="de.tavendo.autobahn">
+      package="io.crossbar.autobahn">
     <application
         android:allowBackup="true"
         android:label="Autobahn Android"
