+++ autobahn/src/main/java/io/crossbar/autobahn/ByteBufferInputStream.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 import java.io.IOException;
 import java.io.InputStream;
