+++ autobahn/src/main/java/io/crossbar/autobahn/Doxygen.java
@@ -1,4 +1,4 @@
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 /*!
 \mainpage
