+++ autobahn/src/main/java/io/crossbar/autobahn/PrefixMap.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 import java.util.HashMap;
