+++ autobahn/src/main/java/io/crossbar/autobahn/Utf8Validator.java
@@ -19,7 +19,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 
 /**
