+++ autobahn/src/main/java/io/crossbar/autobahn/WampConnection.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 import android.os.HandlerThread;
 import android.util.Log;
