+++ autobahn/src/main/java/io/crossbar/autobahn/WampCra.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 public interface WampCra extends Wamp {
