+++ autobahn/src/main/java/io/crossbar/autobahn/WampCraConnection.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 import java.security.InvalidKeyException;
 import java.security.Key;
