+++ autobahn/src/main/java/io/crossbar/autobahn/WampCraPubSubPermission.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 public class WampCraPubSubPermission {
