+++ autobahn/src/main/java/io/crossbar/autobahn/WampOptions.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 public class WampOptions extends WebSocketOptions {
