+++ autobahn/src/main/java/io/crossbar/autobahn/WebSocket.java
@@ -1,4 +1,4 @@
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 public interface WebSocket {
