+++ autobahn/src/main/java/io/crossbar/autobahn/WebSocketConnection.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 import android.os.Handler;
 import android.os.HandlerThread;
@@ -432,7 +432,7 @@ public class WebSocketConnection implements WebSocket {
 
                     if (DEBUG) Log.d(TAG, "WebSockets Close received (" + close.mCode + " - " + close.mReason + ")");
 
-                    final int tavendoCloseCode = (close.mCode == 1000) ? ConnectionHandler.CLOSE_NORMAL : ConnectionHandler.CLOSE_CONNECTION_LOST;
+                    final int crossbarCloseCode = (close.mCode == 1000) ? ConnectionHandler.CLOSE_NORMAL : ConnectionHandler.CLOSE_CONNECTION_LOST;
 
                     if (mActive) {
                         mWriter.forward(new WebSocketMessage.Close(1000));
@@ -445,7 +445,7 @@ public class WebSocketConnection implements WebSocket {
                         }
                     }
 
-                    onClose(tavendoCloseCode, close.mReason);
+                    onClose(crossbarCloseCode, close.mReason);
 
                 } else if (msg.obj instanceof WebSocketMessage.ServerHandshake) {
