+++ autobahn/src/main/java/io/crossbar/autobahn/WebSocketOptions.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 
 /**
