+++ autobahn/src/main/java/io/crossbar/autobahn/WebSocketWriter.java
@@ -16,7 +16,7 @@
  *
  ******************************************************************************/
 
-package de.tavendo.autobahn;
+package io.crossbar.autobahn;
 
 import android.os.Handler;
 import android.os.Looper;
